Add test project

+++ DnmEplusPassword.Web/Models/PlacementAcre.cs
@@ -2,9 +2,9 @@ namespace DnmEplusPassword.Web.Models;
 
 public sealed record PlacementAcre
 {
-    public int Row { get; set; } = ValidRows.First();
-    public int Col { get; set; } = ValidCols.First();
+    public byte Row { get; set; } = ValidRows.First();
+    public byte Col { get; set; } = ValidCols.First();
 
-    public static IReadOnlyList<int> ValidRows { get; } = [1, 2, 3, 4, 5, 6];
-    public static IReadOnlyList<int> ValidCols { get; } = [1, 2, 3, 4, 5];
+    public static IReadOnlyList<byte> ValidRows { get; } = [1, 2, 3, 4, 5, 6];
+    public static IReadOnlyList<byte> ValidCols { get; } = [1, 2, 3, 4, 5];
 }
